% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classSolarAssoc.R
\name{annotate}
\alias{annotate}
\title{Annotate SNPs in association study}
\usage{
annotate(x, ...)
}
\arguments{
\item{x}{An object of class \code{solarAssoc} or a character vector of SNPs.}

\item{...}{Additional arguments passed to \code{annotateSNPs}.}
}
\value{
A data table with annotation results.
}
\description{
The function calls \code{\link{annotateSNPs}} function,
which does the job.
}
\seealso{
\code{\link{annotateSNPs}}
}

