% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/solarBayesAvg.R
\name{solarBaeysAvgPolygenicClass}
\alias{print.solarBaeysAvgPolygenic}
\alias{solarBaeysAvgPolygenicClass}
\alias{summary.solarBaeysAvgPolygenic}
\title{S3 class solarBaeysAvgPolygenic}
\usage{
\method{print}{solarBaeysAvgPolygenic}(x, ...)

\method{summary}{solarBaeysAvgPolygenic}(object, ...)
}
\arguments{
\item{x}{An object of class \code{solarBaeysAvgPolygenic}.}

\item{...}{Additional arguments.}

\item{object}{An object of class \code{solarBaeysAvgPolygenic}.}
}
\description{
S3 class solarBaeysAvgPolygenic
}

