% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{solariusPackage}
\alias{solariusPackage}
\alias{solariusPackage-package}
\title{solarius: an R package wrapper to SOLAR.}
\description{
solarius: an R package wrapper to SOLAR.
}

