% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classSolarPolygenic.R
\name{solarPolygenicClass}
\alias{print.solarPolygenic}
\alias{residuals.solarPolygenic}
\alias{solarPolygenicClass}
\alias{summary.solarPolygenic}
\title{S3 class solarPolygenic.}
\usage{
\method{print}{solarPolygenic}(x, ...)

\method{summary}{solarPolygenic}(object, ...)

\method{residuals}{solarPolygenic}(object, trait = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{solarPolygenic}.}

\item{...}{Additional arguments.}

\item{object}{An object of class \code{solarPolygenic}.}

\item{trait}{Logical argument for \code{residuals} method,
indicating whether values of trait to be returned instead of residuals.
The default value is FALSE.}
}
\description{
S3 class solarPolygenic.
}

