% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classSolarMultipoint.R
\name{solarMultipointClass}
\alias{plot.solarMultipoint}
\alias{print.solarMultipoint}
\alias{solarMultipointClass}
\alias{summary.solarMultipoint}
\title{S3 class solarMultipoint.}
\usage{
\method{print}{solarMultipoint}(x, ...)

\method{plot}{solarMultipoint}(x, pass = 1, ...)

\method{summary}{solarMultipoint}(object, ...)
}
\arguments{
\item{x}{An object of class \code{solarMultipoint}.}

\item{...}{Additional arguments.}

\item{pass}{Integer argument for \code{plot} method,
indicating whether which pass in multi-passs linkage scan to be plotted.
The default value is 1.}

\item{object}{An object of class \code{solarMultipoint}.}
}
\description{
S3 class solarMultipoint.
}

